requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
apache-airflow-providers-google
apache-airflow-providers-postgres
apache-airflow-providers-http
pandas
requests
google-cloud-bigquery
google-cloud-storage
